chore: AA refactor
