+++ src/components/playlists/Playlists.tsx
@@ -18,7 +18,7 @@ import useAlert from '../dialogs/useAlert';
 import ShuffleAllButton from '@components/playlists/ShuffleAllButton';
 import TimerButton from '@components/playlists/TimerButton';
 import PlaylistItem from '@components/playlists/PlaylistItem';
-import usePlaylistAA from '@hooks/usePlaylistAA';
+import usePlaylistBrowseTree from '@/hooks/usePlaylistBrowseTree';
 import { BottomTabRouteIcons as RouteIcons } from '@enums/BottomTab';
 
 export default () => {
@@ -35,7 +35,7 @@ export default () => {
   const setPlaylistIds = useNoxSetting(state => state.setPlaylistIds);
   const scroll = useNoxSetting(state => state.incSongListScrollCounter);
   const setRoute = useNoxSetting(state => state.setBottomTabRoute);
-  const { removePlaylist } = usePlaylistAA();
+  const { removePlaylist } = usePlaylistBrowseTree();
   const { TwoWayAlert } = useAlert();
   // HACK: I know its bad! But somehow this hook isnt updating in its own
   // useEffects...
