+++ src/hooks/useAndroidAutoPermission.ts
