+++ src/hooks/usePlaylistBrowseTree.ts
@@ -1,10 +1,13 @@
 import { useNoxSetting } from '@stores/useApp';
-import usePlayback from './usePlayback';
+import { useAndroidAuto } from './usePlaybackAA';
 
+/**
+ * wrapper of playlist operations with browsetree updates.
+ */
 export default () => {
   const addPlaylistStore = useNoxSetting(state => state.addPlaylist);
   const removePlaylistStore = useNoxSetting(state => state.removePlaylist);
-  const { buildBrowseTree } = usePlayback();
+  const { buildBrowseTree } = useAndroidAuto();
 
   const addPlaylist = (playlist: NoxMedia.Playlist) => {
     addPlaylistStore(playlist);
